Udacity-Data-Wrangling-WeRateDogs
What is WERATEDOGS™?

WeRateDogs IS a Twitter Account That Ranks people's dogs with a funny dog comment. The account was founded in 2015 by Matt Nelson, a college student, and gained international media coverage both for its success and for the attention it paid to the copyright law on social media, when Twitter was suspended for breaching the above-mentioned rules.
WeRateDogs has over 6 million followers and has received international media
Coverage. One of those coverage were about the quote “they’re good dogs Brent”,This was an exchange in which WeRateDogs shut down a person having an issue with its rating system in humorous ways.